Update store page
The items.json file was created to represent the items sold in the
store.
The StoreItem.tsx file was created to return the styled cards for
each of the items in the json file.
The formatCurrency.ts file was created to create a function that formats
the given number into european currency.
Store.tsx file was modified to include these additions.