Create ShoppingCartContext.tsx, modify image files
ShoppingCartContext.tsx was created for storing functions that add and
remove an item from the shopping cart as well as increase and decrease
the quantity of items in the cart.
Images were moved to the separate folder and links to these images were
changed in the json file.
Navbar and StoreItems components were modified to reflect the above
changes.